switch from esbuild to tsc
