+++ .gitignore
@@ -4,5 +4,4 @@
 node_modules
 coverage
 .nyc_output
-cjs
-esm
+dist
