+++ .xo-config.json
@@ -13,6 +13,7 @@
     "import/no-anonymous-default-export": "off",
     "@typescript-eslint/comma-dangle": "off",
     "@typescript-eslint/ban-types": "off",
-    "@typescript-eslint/no-unsafe-call": "off"
+    "@typescript-eslint/no-unsafe-call": "off",
+    "unicorn/prefer-module": "off"
   }
 }
