+++ package.json
@@ -31,8 +31,6 @@
     "@typescript-eslint/parser": "^5.56.0",
     "codecov": "^3.8.2",
     "cross-env": "^7.0.3",
-    "esbuild": "^0.17.14",
-    "esbuild-jest": "^0.5.0",
     "eslint": "^8.36.0",
     "eslint-config-xo": "^0.43.1",
     "eslint-config-xo-lass": "^2.0.1",
@@ -93,13 +91,9 @@
     "nyc": "cross-env NODE_ENV=test nyc ava",
     "test": "jest",
     "test-coverage": "yarn run lint && yarn run nyc",
-    "build": "npx esbuild --bundle src/index.ts --outdir=cjs --platform=node --format=cjs && echo '{\"type\": \"commonjs\"}' > cjs/package.json && npx esbuild --bundle src/index.ts --outdir=esm --platform=neutral --format=esm && echo '{\"type\": \"module\"}' > esm/package.json"
+    "build": "npx tsc -p tsconfig.build.json --outDir dist/ && echo '{\"type\": \"commonjs\"}' > dist/package.json"
   },
-  "exports": {
-    "require": "./cjs/index.js",
-    "import": "./esm/index.js"
-  },
-  "main": "./cjs/index.js",
+  "main": "./dist/index.js",
   "types": "./src",
   "overrides": {
     "decamelize": "4.0.0"
