+++ test/sensitive-paths.test.ts
@@ -1,4 +1,4 @@
-import { DiffEngine as HR } from '../src/index';
+import HR from '../src/index';
 
 describe('sensitive-paths', () => {
   let hr: Function;
